ENH: Add support to overrule configuration parameters via command line arguments - fix (#154)

+++ orthoseg/load_images.py
@@ -191,7 +191,8 @@ def load_images(
         )
         raise Exception(message) from ex
     finally:
-        shutil.rmtree(conf.tmp_dir, ignore_errors=True)
+        if conf.tmp_dir is not None:
+            shutil.rmtree(conf.tmp_dir, ignore_errors=True)
 
 
 def main():
